+++ cardinal-kit-web/src/components/tables/altTable.vue
@@ -12,30 +12,14 @@
        <slot name="t-row" />
       </tbody>
     </table>
-    <div v-if="pagination" class="pagination">
-      <div class="flex mr-2 align-center pagination-child">
-        <label class="mr-1 pagination-label">Items per page</label>
-        <select class="select-table" name="limit" v-model="limit">    
-          <option v-for="(value, index) in paginationOptions.limit" :key="index" :value="value">{{value}}</option>
-        </select>
-      </div>
-      <div class="flex mr-2 pagination-child">
-        <p>{{range}} of {{paginationOptions.total}}</p>
-      </div>
-      <div class="flex align-center pagination-child">
-        <select class="select-table" name="page" v-model="currentPage">
-          <option v-for="(page, index) in pages" :key="index" :value="page">{{page}}</option>
-        </select>
-      </div>
-    </div>
-    <pre>{{limit}} | {{currentPage}}</pre>
+    <Pagination v-if="pagination" :pagination="paginationOptions" @onChangePagination="handleChangePagination"/>
   </div>
 </template>
 
 <script>
-import { computed, ref, watch } from 'vue';
-
+import Pagination from './Pagination'
 export default {
+  components: { Pagination },
   props: {
     columns:{
       type: Array,
@@ -50,7 +34,7 @@ export default {
     },
     pagination: {
       type: Boolean,
-      default: true
+      default: false
     },
     paginationOptions: {
       type: Object,
@@ -61,37 +45,9 @@ export default {
       }
     }
   },
-  setup(props, ctx) {
-    const limit = ref(10);
-    const currentPage = ref(props.paginationOptions.current || 1);
-    const pages = computed(() => Array.from({length: getCountPage() }, (_, i) => i+1));
-    const range = computed(() => calculateRange(currentPage.value, limit.value, props.paginationOptions.total));
-    
-    const getCountPage = () => (Math.ceil(props.paginationOptions.total / limit.value));
-
-    function calculateRange(page, limit, total) {
-      console.log('total', total);
-      const initRange = ((page*limit) - limit + 1);
-      const finalRange = isLastPage(page) ? total : limit*page;
-      return `${initRange}-${finalRange}`
-    }
-
-    const isLastPage = (page) => (pages.value.length === page);
-
-    watch([limit, currentPage],() => {
-      const values = {
-        currentPage: currentPage.value,
-        limit: limit.value
-      }
-      ctx.emit('pagination', { ...values })
-    });
-
-
-    return {
-      limit,
-      pages,
-      currentPage,
-      range
+  methods: {
+    handleChangePagination (params) {
+      this.$emit('onPagination', params);
     }
   }
 }
@@ -151,42 +107,6 @@ export default {
       line-height: 1.25rem;
       padding: 1.25rem;
       border-bottom-width: 1px;
-    }
-  }
-
-  .select-table {
-    padding: 3px 1rem;
-    border: none;
-    border-bottom: solid;
-    border-radius: 5px;
-    border-color: $light-grey;
-    outline: none;
-  }
-  
-  .pagination {
-    padding: .5rem 1rem;
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-    border-top: 1px solid;
-    font-size: 10pt;
-    border-color: $light-grey;
-
-    @media (max-width: 520px) {
-      &-child {
-        justify-content: center;
-        width: 100%;
-        padding: .3rem 0;
-        border-color: $light-grey;
-      }
-
-      &-label {
-        display: none;
-      }
-
-      .select-table {
-       width: 100%; 
-      }
     }
   }
 }
